Design Pattern
Implementing a Non-Preemptive Process Management Application. (Imagine that this is a multiprogramming environment, therefore scheduling is putting processes in an order to be executed sequentially in CPU, there are no parallel processes. One single CPU executes
processes one after the other. My assignment asks to follow the rules defined on the pdf, not to the real world OS rules.)